+++ nuxt/assets/styles/main.scss
@@ -5,7 +5,10 @@ body {
   font-family: "Inter", "sans-serif";
 }
 
-.main-container {
-  max-width: 1000px;
-  margin: 0 auto;
+.bg-drink {
+  background-image: url("~/assets/img/keyvisual-pc.jpg");
+  background-color: rgba(#ffffff, 0.35);
+  background-blend-mode: overlay;
+  background-size: cover;
+  height: calc(100dvh - 72px);
 }
