+++ nuxt/components/AppFooter.vue
@@ -1,0 +1,14 @@
+<template>
+  <footer class="footer items-center p-4 bg-neutral text-neutral-content">
+    <div
+      class="w-full max-w-screen-lg mx-auto flex justify-between items-center"
+    >
+      <aside class="items-center pl-3">
+        <p>Copyright © 2024 - All right reserved by @shun0608</p>
+      </aside>
+      <a class="p-2" href="https://github.com/shun0608/drinko-app"
+        ><i class="fab fa-github-square text-xl"></i
+      ></a>
+    </div>
+  </footer>
+</template>
